requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.0.1
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
whitenoise==6.6.0
